+++ objective-c/web-page-tutorial/UsingViews2/UsingViews2/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  UsingViews2
+//
+//  Created by Juan Gallardo on 1/15/13.
+//  Copyright (c) 2013 __MyCompanyName__. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@end