+++ objective-c/web-page-tutorial/UsingViews2/UsingViews2/UsingViews2AppDelegate.h
@@ -1,0 +1,19 @@
+//
+//  UsingViews2AppDelegate.h
+//  UsingViews2
+//
+//  Created by Juan Gallardo on 1/15/13.
+//  Copyright (c) 2013 __MyCompanyName__. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@class UsingViews2ViewController;
+
+@interface UsingViews2AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@property (strong, nonatomic) UsingViews2ViewController *viewController;
+
+@end