+++ objective-c/web-page-tutorial/UsingViews2/UsingViews2/UsingViews2ViewController.h
@@ -1,0 +1,18 @@
+//
+//  UsingViews2ViewController.h
+//  UsingViews2
+//
+//  Created by Juan Gallardo on 1/15/13.
+//  Copyright (c) 2013 __MyCompanyName__. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface UsingViews2ViewController : UIViewController 
+{
+	IBOutlet UIWebView *webView;
+}
+
+@property (nonatomic, retain) UIWebView *webView;
+
+@end