+++ objective-c/web-page-tutorial/UsingViews2/UsingViews2/main.m
@@ -1,0 +1,18 @@
+//
+//  main.m
+//  UsingViews2
+//
+//  Created by Juan Gallardo on 1/15/13.
+//  Copyright (c) 2013 __MyCompanyName__. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+#import "UsingViews2AppDelegate.h"
+
+int main(int argc, char *argv[])
+{
+	@autoreleasepool {
+	    return UIApplicationMain(argc, argv, nil, NSStringFromClass([UsingViews2AppDelegate class]));
+	}
+}