+++ HISTORY.md
@@ -1,6 +1,31 @@
 HISTORY
 ==================
 
+__2013-09-01__ Juan Gallardo <j@jgallardo.me> <br />
++ phonegap/contacts-tutorial/ .... / : Added photos.html and photos.js which together allow you to upload an image from the device or take and store a picture.
+
+__2013-05-01__ Juan Gallardo <j@jgallardo.me> <br />
++ phonegap/contacts : Added the function to create a new contact. Created form.html and contacts.js.  
+
+__2012-30-18__ Juan Gallardo <j@jgallardo.me> <br />
++ ruby/createGallery.rb : Finished tutorial on how to generate a gallery for images.
+
+__2012-23-18__ Juan Gallardo <j@jgallardo.me> <br />
++ phonegap/using-gps-and-displaying-a-position-on-a-map : Finished tutorial on how to integrate the Google Maps API into a phonegap application so that the user can view their location. 
++ phonegap/using-gps-and-displaying-a-position-on-a-map/GPS tutorial/www/compass.html : Finised a tutorial on how to add compass functionality. Added the compass.html, compass.js, needle.png, compass.png, and the jqueryrotate plugin to the previous file of using-and-displaying-a-position-on-a-map.
++ phonegap/using-gps-and-displaying-a-position-on-a-map/GPS tutorial/www/accelerometer.html : Finised a tutorial on how to add the accelerometer to detect motion. Added the accelerometer.html, accelerometer.js, to the previous file of using-and-displaying-a-position-on-a-map.  
++ phonegap/contacts-tutorial : Added this project to seperate files that could be used for a contact management app, instead of continuing to add files to the previous project as suggested by the book.
++ phonegap/contacts-tutorial/ .... /list.html : Finished the tutorial on how to display Table-View data. Added list.html, 
++ phonegap/contacts-tutorial/ .... /contacts.html : Finished the tutorial on how to pull contacts from the device and then display them in the list view. Added contacts.html and contacts.js. 
+
+
+__2012-22-18__ Juan Gallardo <j@jgallardo.me> <br />
++ phonegap/persistent-navigation-tutorial : Finished the tutorial on building the template for a PhoneGap application that has a persistent footer.
++ phonegap/network-detection.js : Finished tutorial on how to add functions in JavaScript to detect a network connection. Modified "common.js" from another file and added functions. 
++ phonegap/network-status-change.js : Finished tutorial on adding JavaScript functions that detect when a change in network connection is detected. 
++ phonegap/callback-function-tutorial : Finished the tutorial on implementing a callback with jQuery
++ phonegap/background-foreground-dection.js : Finished tutorial to add code that will enable the application to detect when the app moves to the background or foreground. 
+
 __2012-12-18__ Juan Gallardo <j@jgallardo.me> <br />
 + phonegap/determining-whether-the-device-is-ready : Finished the tutorial 
 + phonegap/retrieve.js : Finished tutorial on writing a JavaScript file that enables you to add functionality that is available only on a certain device or platform.